Setup script to check example name and directory
The script is setup to take in, as its first parameter, the name of the
example to build. This is then checked against the valid examples and
matched with the associated directory. The directory is then checked for
in the current path - if not, it's an indication that the script was not
sourced from the project root.