Update script with build instructions for mock example
This commit saw a few major changes, including a correction to the
target board variable. The mount point is now passed in as an argument
to the script as there's an issue with flashing with mbed-tools; please
refer to issue #282 on the mbed-tools repository for more information.
As a result, the binary is flashed manually.