Update script with build instructions for MCUboot example
The script has many caveats that are addressed as a comment on the
associated issue (or pull request). These are mainly as a result of
dependency conflicts between mbed-tools, mbed-os, and pyocd. The updated
script follows the build instructions as listed in the documentation
accompanying this example (with additional checks).