Add "fail" function to utils
The function prints an error message to stderr and exits or returns with
a code of 1 based on the fail mode, which is passed as an input. If the
file is being sourced, then "return" would be passed in; else "exit"
would be passed in. This function invokes the say function created
earlier with the error formatting mode.