Update parse options to parse space-separated options
Previously, the parse_options function required that the option be
followed by an '=' symbol and the value (i.e -o=value or
--option=value). The usage instructions separate the option and the
value by a space and the function was modified to meet this requirement.

The following stack overflow post proved quite useful in understanding
how options are parsed - in fact, the "Bash Equals-Separated" part of
the accepted answer was adapted for the intial implementation.

https://stackoverflow.com/a/14203146