Fix indefinite looping bug in options parsing function
The parsing function would loop indefinintely if there was no value
associated with an option. There is still a need to check if the default
value of the options have been cleared, in which case the prior default
must be restored or an error must be displayed indicating that the
option was not followed by a value.