Add dummy example build functions
This change introduces the build_example function that calls a dummy
build function in the corresponding example's shell scripts (mock.sh and
mcuboot.sh).  It also passes the valid arguments to this build function.
The next few commits will see work on these functions, starting with the
build in mock.sh.