Update mcuboot_build to handle changes to mbed_app.json
This is quite a significant commit - the script now prompts the user on
whether it should auto-update the application/mbed_app.json file; this
serves well for testing purposes where the "yes" would be piped into the
script. Also, part of the script was adapted from the prior
implementation in build.sh.

Note that this introduces a new dependency "jq" that's not installed
through pip - rather, it requires some manual intervention on the
end-user's behalf.